Fix from codereview
