+++ src/AppModel/NetDaemon.AppModel/Internal/Config/IConfigurationBinding.cs
@@ -1,0 +1,11 @@
+using Microsoft.Extensions.Configuration;
+
+namespace NetDaemon.AppModel.Internal.Config;
+
+/// <summary>
+/// Interface for configuration binding to object
+/// </summary>
+internal interface IConfigurationBinding
+{
+    T? ToObject<T>(IConfiguration configuration);
+}
